+++ src/Component/JsonSchema/JsonSchema/Normalizer/JsonSchemaNormalizer.php
@@ -69,39 +69,51 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
         }
         if (\array_key_exists('$id', $data) && $data['$id'] !== null) {
             $object->setDollarId($data['$id']);
+            unset($data['$id']);
         }
         elseif (\array_key_exists('$id', $data) && $data['$id'] === null) {
             $object->setDollarId(null);
+            unset($data['$id']);
         }
         if (\array_key_exists('$schema', $data) && $data['$schema'] !== null) {
             $object->setDollarSchema($data['$schema']);
+            unset($data['$schema']);
         }
         elseif (\array_key_exists('$schema', $data) && $data['$schema'] === null) {
             $object->setDollarSchema(null);
+            unset($data['$schema']);
         }
         if (\array_key_exists('$ref', $data) && $data['$ref'] !== null) {
             $object->setDollarRef($data['$ref']);
+            unset($data['$ref']);
         }
         elseif (\array_key_exists('$ref', $data) && $data['$ref'] === null) {
             $object->setDollarRef(null);
+            unset($data['$ref']);
         }
         if (\array_key_exists('$anchor', $data) && $data['$anchor'] !== null) {
             $object->setDollarAnchor($data['$anchor']);
+            unset($data['$anchor']);
         }
         elseif (\array_key_exists('$anchor', $data) && $data['$anchor'] === null) {
             $object->setDollarAnchor(null);
+            unset($data['$anchor']);
         }
         if (\array_key_exists('$dynamicRef', $data) && $data['$dynamicRef'] !== null) {
             $object->setDollarDynamicRef($data['$dynamicRef']);
+            unset($data['$dynamicRef']);
         }
         elseif (\array_key_exists('$dynamicRef', $data) && $data['$dynamicRef'] === null) {
             $object->setDollarDynamicRef(null);
+            unset($data['$dynamicRef']);
         }
         if (\array_key_exists('$dynamicAnchor', $data) && $data['$dynamicAnchor'] !== null) {
             $object->setDollarDynamicAnchor($data['$dynamicAnchor']);
+            unset($data['$dynamicAnchor']);
         }
         elseif (\array_key_exists('$dynamicAnchor', $data) && $data['$dynamicAnchor'] === null) {
             $object->setDollarDynamicAnchor(null);
+            unset($data['$dynamicAnchor']);
         }
         if (\array_key_exists('$vocabulary', $data) && $data['$vocabulary'] !== null) {
             $values = new \Jane\Component\JsonSchema\JsonSchema\Runtime\JsonObject();
@@ -109,15 +121,19 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $values[$key] = $value;
             }
             $object->setDollarVocabulary($values);
+            unset($data['$vocabulary']);
         }
         elseif (\array_key_exists('$vocabulary', $data) && $data['$vocabulary'] === null) {
             $object->setDollarVocabulary(null);
+            unset($data['$vocabulary']);
         }
         if (\array_key_exists('$comment', $data) && $data['$comment'] !== null) {
             $object->setDollarComment($data['$comment']);
+            unset($data['$comment']);
         }
         elseif (\array_key_exists('$comment', $data) && $data['$comment'] === null) {
             $object->setDollarComment(null);
+            unset($data['$comment']);
         }
         if (\array_key_exists('$defs', $data) && $data['$defs'] !== null) {
             $values_1 = new \Jane\Component\JsonSchema\JsonSchema\Runtime\JsonObject();
@@ -131,9 +147,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $values_1[$key_1] = $value_2;
             }
             $object->setDollarDefs($values_1);
+            unset($data['$defs']);
         }
         elseif (\array_key_exists('$defs', $data) && $data['$defs'] === null) {
             $object->setDollarDefs(null);
+            unset($data['$defs']);
         }
         if (\array_key_exists('prefixItems', $data) && $data['prefixItems'] !== null) {
             $values_2 = [];
@@ -147,9 +165,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $values_2[] = $value_4;
             }
             $object->setPrefixItems($values_2);
+            unset($data['prefixItems']);
         }
         elseif (\array_key_exists('prefixItems', $data) && $data['prefixItems'] === null) {
             $object->setPrefixItems(null);
+            unset($data['prefixItems']);
         }
         if (\array_key_exists('items', $data) && $data['items'] !== null) {
             $value_5 = $data['items'];
@@ -171,9 +191,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $value_5 = $values_3;
             }
             $object->setItems($value_5);
+            unset($data['items']);
         }
         elseif (\array_key_exists('items', $data) && $data['items'] === null) {
             $object->setItems(null);
+            unset($data['items']);
         }
         if (\array_key_exists('contains', $data) && $data['contains'] !== null) {
             $value_8 = $data['contains'];
@@ -183,9 +205,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $value_8 = $data['contains'];
             }
             $object->setContains($value_8);
+            unset($data['contains']);
         }
         elseif (\array_key_exists('contains', $data) && $data['contains'] === null) {
             $object->setContains(null);
+            unset($data['contains']);
         }
         if (\array_key_exists('additionalProperties', $data) && $data['additionalProperties'] !== null) {
             $value_9 = $data['additionalProperties'];
@@ -195,9 +219,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $value_9 = $data['additionalProperties'];
             }
             $object->setAdditionalProperties($value_9);
+            unset($data['additionalProperties']);
         }
         elseif (\array_key_exists('additionalProperties', $data) && $data['additionalProperties'] === null) {
             $object->setAdditionalProperties(null);
+            unset($data['additionalProperties']);
         }
         if (\array_key_exists('properties', $data) && $data['properties'] !== null) {
             $values_4 = new \Jane\Component\JsonSchema\JsonSchema\Runtime\JsonObject();
@@ -211,9 +237,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $values_4[$key_2] = $value_11;
             }
             $object->setProperties($values_4);
+            unset($data['properties']);
         }
         elseif (\array_key_exists('properties', $data) && $data['properties'] === null) {
             $object->setProperties(null);
+            unset($data['properties']);
         }
         if (\array_key_exists('patternProperties', $data) && $data['patternProperties'] !== null) {
             $values_5 = new \Jane\Component\JsonSchema\JsonSchema\Runtime\JsonObject();
@@ -227,9 +255,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $values_5[$key_3] = $value_13;
             }
             $object->setPatternProperties($values_5);
+            unset($data['patternProperties']);
         }
         elseif (\array_key_exists('patternProperties', $data) && $data['patternProperties'] === null) {
             $object->setPatternProperties(null);
+            unset($data['patternProperties']);
         }
         if (\array_key_exists('dependentSchemas', $data) && $data['dependentSchemas'] !== null) {
             $values_6 = new \Jane\Component\JsonSchema\JsonSchema\Runtime\JsonObject();
@@ -243,9 +273,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $values_6[$key_4] = $value_15;
             }
             $object->setDependentSchemas($values_6);
+            unset($data['dependentSchemas']);
         }
         elseif (\array_key_exists('dependentSchemas', $data) && $data['dependentSchemas'] === null) {
             $object->setDependentSchemas(null);
+            unset($data['dependentSchemas']);
         }
         if (\array_key_exists('propertyNames', $data) && $data['propertyNames'] !== null) {
             $value_16 = $data['propertyNames'];
@@ -255,9 +287,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $value_16 = $data['propertyNames'];
             }
             $object->setPropertyNames($value_16);
+            unset($data['propertyNames']);
         }
         elseif (\array_key_exists('propertyNames', $data) && $data['propertyNames'] === null) {
             $object->setPropertyNames(null);
+            unset($data['propertyNames']);
         }
         if (\array_key_exists('if', $data) && $data['if'] !== null) {
             $value_17 = $data['if'];
@@ -267,9 +301,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $value_17 = $data['if'];
             }
             $object->setIf($value_17);
+            unset($data['if']);
         }
         elseif (\array_key_exists('if', $data) && $data['if'] === null) {
             $object->setIf(null);
+            unset($data['if']);
         }
         if (\array_key_exists('then', $data) && $data['then'] !== null) {
             $value_18 = $data['then'];
@@ -279,9 +315,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $value_18 = $data['then'];
             }
             $object->setThen($value_18);
+            unset($data['then']);
         }
         elseif (\array_key_exists('then', $data) && $data['then'] === null) {
             $object->setThen(null);
+            unset($data['then']);
         }
         if (\array_key_exists('else', $data) && $data['else'] !== null) {
             $value_19 = $data['else'];
@@ -291,9 +329,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $value_19 = $data['else'];
             }
             $object->setElse($value_19);
+            unset($data['else']);
         }
         elseif (\array_key_exists('else', $data) && $data['else'] === null) {
             $object->setElse(null);
+            unset($data['else']);
         }
         if (\array_key_exists('allOf', $data) && $data['allOf'] !== null) {
             $values_7 = [];
@@ -307,9 +347,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $values_7[] = $value_21;
             }
             $object->setAllOf($values_7);
+            unset($data['allOf']);
         }
         elseif (\array_key_exists('allOf', $data) && $data['allOf'] === null) {
             $object->setAllOf(null);
+            unset($data['allOf']);
         }
         if (\array_key_exists('anyOf', $data) && $data['anyOf'] !== null) {
             $values_8 = [];
@@ -323,9 +365,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $values_8[] = $value_23;
             }
             $object->setAnyOf($values_8);
+            unset($data['anyOf']);
         }
         elseif (\array_key_exists('anyOf', $data) && $data['anyOf'] === null) {
             $object->setAnyOf(null);
+            unset($data['anyOf']);
         }
         if (\array_key_exists('oneOf', $data) && $data['oneOf'] !== null) {
             $values_9 = [];
@@ -339,9 +383,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $values_9[] = $value_25;
             }
             $object->setOneOf($values_9);
+            unset($data['oneOf']);
         }
         elseif (\array_key_exists('oneOf', $data) && $data['oneOf'] === null) {
             $object->setOneOf(null);
+            unset($data['oneOf']);
         }
         if (\array_key_exists('not', $data) && $data['not'] !== null) {
             $value_26 = $data['not'];
@@ -351,9 +397,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $value_26 = $data['not'];
             }
             $object->setNot($value_26);
+            unset($data['not']);
         }
         elseif (\array_key_exists('not', $data) && $data['not'] === null) {
             $object->setNot(null);
+            unset($data['not']);
         }
         if (\array_key_exists('unevaluatedItems', $data) && $data['unevaluatedItems'] !== null) {
             $value_27 = $data['unevaluatedItems'];
@@ -363,9 +411,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $value_27 = $data['unevaluatedItems'];
             }
             $object->setUnevaluatedItems($value_27);
+            unset($data['unevaluatedItems']);
         }
         elseif (\array_key_exists('unevaluatedItems', $data) && $data['unevaluatedItems'] === null) {
             $object->setUnevaluatedItems(null);
+            unset($data['unevaluatedItems']);
         }
         if (\array_key_exists('unevaluatedProperties', $data) && $data['unevaluatedProperties'] !== null) {
             $value_28 = $data['unevaluatedProperties'];
@@ -375,9 +425,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $value_28 = $data['unevaluatedProperties'];
             }
             $object->setUnevaluatedProperties($value_28);
+            unset($data['unevaluatedProperties']);
         }
         elseif (\array_key_exists('unevaluatedProperties', $data) && $data['unevaluatedProperties'] === null) {
             $object->setUnevaluatedProperties(null);
+            unset($data['unevaluatedProperties']);
         }
         if (\array_key_exists('type', $data) && $data['type'] !== null) {
             $value_29 = $data['type'];
@@ -391,15 +443,19 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $value_29 = $data['type'];
             }
             $object->setType($value_29);
+            unset($data['type']);
         }
         elseif (\array_key_exists('type', $data) && $data['type'] === null) {
             $object->setType(null);
+            unset($data['type']);
         }
         if (\array_key_exists('const', $data) && $data['const'] !== null) {
             $object->setConst($data['const']);
+            unset($data['const']);
         }
         elseif (\array_key_exists('const', $data) && $data['const'] === null) {
             $object->setConst(null);
+            unset($data['const']);
         }
         if (\array_key_exists('enum', $data) && $data['enum'] !== null) {
             $values_11 = [];
@@ -407,99 +463,131 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $values_11[] = $value_31;
             }
             $object->setEnum($values_11);
+            unset($data['enum']);
         }
         elseif (\array_key_exists('enum', $data) && $data['enum'] === null) {
             $object->setEnum(null);
+            unset($data['enum']);
         }
         if (\array_key_exists('multipleOf', $data) && $data['multipleOf'] !== null) {
             $object->setMultipleOf($data['multipleOf']);
+            unset($data['multipleOf']);
         }
         elseif (\array_key_exists('multipleOf', $data) && $data['multipleOf'] === null) {
             $object->setMultipleOf(null);
+            unset($data['multipleOf']);
         }
         if (\array_key_exists('maximum', $data) && $data['maximum'] !== null) {
             $object->setMaximum($data['maximum']);
+            unset($data['maximum']);
         }
         elseif (\array_key_exists('maximum', $data) && $data['maximum'] === null) {
             $object->setMaximum(null);
+            unset($data['maximum']);
         }
         if (\array_key_exists('exclusiveMaximum', $data) && $data['exclusiveMaximum'] !== null) {
             $object->setExclusiveMaximum($data['exclusiveMaximum']);
+            unset($data['exclusiveMaximum']);
         }
         elseif (\array_key_exists('exclusiveMaximum', $data) && $data['exclusiveMaximum'] === null) {
             $object->setExclusiveMaximum(null);
+            unset($data['exclusiveMaximum']);
         }
         if (\array_key_exists('minimum', $data) && $data['minimum'] !== null) {
             $object->setMinimum($data['minimum']);
+            unset($data['minimum']);
         }
         elseif (\array_key_exists('minimum', $data) && $data['minimum'] === null) {
             $object->setMinimum(null);
+            unset($data['minimum']);
         }
         if (\array_key_exists('exclusiveMinimum', $data) && $data['exclusiveMinimum'] !== null) {
             $object->setExclusiveMinimum($data['exclusiveMinimum']);
+            unset($data['exclusiveMinimum']);
         }
         elseif (\array_key_exists('exclusiveMinimum', $data) && $data['exclusiveMinimum'] === null) {
             $object->setExclusiveMinimum(null);
+            unset($data['exclusiveMinimum']);
         }
         if (\array_key_exists('maxLength', $data) && $data['maxLength'] !== null) {
             $object->setMaxLength($data['maxLength']);
+            unset($data['maxLength']);
         }
         elseif (\array_key_exists('maxLength', $data) && $data['maxLength'] === null) {
             $object->setMaxLength(null);
+            unset($data['maxLength']);
         }
         if (\array_key_exists('minLength', $data) && $data['minLength'] !== null) {
             $object->setMinLength($data['minLength']);
+            unset($data['minLength']);
         }
         elseif (\array_key_exists('minLength', $data) && $data['minLength'] === null) {
             $object->setMinLength(null);
+            unset($data['minLength']);
         }
         if (\array_key_exists('pattern', $data) && $data['pattern'] !== null) {
             $object->setPattern($data['pattern']);
+            unset($data['pattern']);
         }
         elseif (\array_key_exists('pattern', $data) && $data['pattern'] === null) {
             $object->setPattern(null);
+            unset($data['pattern']);
         }
         if (\array_key_exists('maxItems', $data) && $data['maxItems'] !== null) {
             $object->setMaxItems($data['maxItems']);
+            unset($data['maxItems']);
         }
         elseif (\array_key_exists('maxItems', $data) && $data['maxItems'] === null) {
             $object->setMaxItems(null);
+            unset($data['maxItems']);
         }
         if (\array_key_exists('minItems', $data) && $data['minItems'] !== null) {
             $object->setMinItems($data['minItems']);
+            unset($data['minItems']);
         }
         elseif (\array_key_exists('minItems', $data) && $data['minItems'] === null) {
             $object->setMinItems(null);
+            unset($data['minItems']);
         }
         if (\array_key_exists('uniqueItems', $data) && $data['uniqueItems'] !== null) {
             $object->setUniqueItems($data['uniqueItems']);
+            unset($data['uniqueItems']);
         }
         elseif (\array_key_exists('uniqueItems', $data) && $data['uniqueItems'] === null) {
             $object->setUniqueItems(null);
+            unset($data['uniqueItems']);
         }
         if (\array_key_exists('maxContains', $data) && $data['maxContains'] !== null) {
             $object->setMaxContains($data['maxContains']);
+            unset($data['maxContains']);
         }
         elseif (\array_key_exists('maxContains', $data) && $data['maxContains'] === null) {
             $object->setMaxContains(null);
+            unset($data['maxContains']);
         }
         if (\array_key_exists('minContains', $data) && $data['minContains'] !== null) {
             $object->setMinContains($data['minContains']);
+            unset($data['minContains']);
         }
         elseif (\array_key_exists('minContains', $data) && $data['minContains'] === null) {
             $object->setMinContains(null);
+            unset($data['minContains']);
         }
         if (\array_key_exists('maxProperties', $data) && $data['maxProperties'] !== null) {
             $object->setMaxProperties($data['maxProperties']);
+            unset($data['maxProperties']);
         }
         elseif (\array_key_exists('maxProperties', $data) && $data['maxProperties'] === null) {
             $object->setMaxProperties(null);
+            unset($data['maxProperties']);
         }
         if (\array_key_exists('minProperties', $data) && $data['minProperties'] !== null) {
             $object->setMinProperties($data['minProperties']);
+            unset($data['minProperties']);
         }
         elseif (\array_key_exists('minProperties', $data) && $data['minProperties'] === null) {
             $object->setMinProperties(null);
+            unset($data['minProperties']);
         }
         if (\array_key_exists('required', $data) && $data['required'] !== null) {
             $values_12 = [];
@@ -507,9 +595,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $values_12[] = $value_32;
             }
             $object->setRequired($values_12);
+            unset($data['required']);
         }
         elseif (\array_key_exists('required', $data) && $data['required'] === null) {
             $object->setRequired(null);
+            unset($data['required']);
         }
         if (\array_key_exists('dependentRequired', $data) && $data['dependentRequired'] !== null) {
             $values_13 = new \Jane\Component\JsonSchema\JsonSchema\Runtime\JsonObject();
@@ -521,45 +611,59 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $values_13[$key_5] = $values_14;
             }
             $object->setDependentRequired($values_13);
+            unset($data['dependentRequired']);
         }
         elseif (\array_key_exists('dependentRequired', $data) && $data['dependentRequired'] === null) {
             $object->setDependentRequired(null);
+            unset($data['dependentRequired']);
         }
         if (\array_key_exists('title', $data) && $data['title'] !== null) {
             $object->setTitle($data['title']);
+            unset($data['title']);
         }
         elseif (\array_key_exists('title', $data) && $data['title'] === null) {
             $object->setTitle(null);
+            unset($data['title']);
         }
         if (\array_key_exists('description', $data) && $data['description'] !== null) {
             $object->setDescription($data['description']);
+            unset($data['description']);
         }
         elseif (\array_key_exists('description', $data) && $data['description'] === null) {
             $object->setDescription(null);
+            unset($data['description']);
         }
         if (\array_key_exists('default', $data) && $data['default'] !== null) {
             $object->setDefault($data['default']);
+            unset($data['default']);
         }
         elseif (\array_key_exists('default', $data) && $data['default'] === null) {
             $object->setDefault(null);
+            unset($data['default']);
         }
         if (\array_key_exists('deprecated', $data) && $data['deprecated'] !== null) {
             $object->setDeprecated($data['deprecated']);
+            unset($data['deprecated']);
         }
         elseif (\array_key_exists('deprecated', $data) && $data['deprecated'] === null) {
             $object->setDeprecated(null);
+            unset($data['deprecated']);
         }
         if (\array_key_exists('readOnly', $data) && $data['readOnly'] !== null) {
             $object->setReadOnly($data['readOnly']);
+            unset($data['readOnly']);
         }
         elseif (\array_key_exists('readOnly', $data) && $data['readOnly'] === null) {
             $object->setReadOnly(null);
+            unset($data['readOnly']);
         }
         if (\array_key_exists('writeOnly', $data) && $data['writeOnly'] !== null) {
             $object->setWriteOnly($data['writeOnly']);
+            unset($data['writeOnly']);
         }
         elseif (\array_key_exists('writeOnly', $data) && $data['writeOnly'] === null) {
             $object->setWriteOnly(null);
+            unset($data['writeOnly']);
         }
         if (\array_key_exists('examples', $data) && $data['examples'] !== null) {
             $values_15 = [];
@@ -567,27 +671,35 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $values_15[] = $value_35;
             }
             $object->setExamples($values_15);
+            unset($data['examples']);
         }
         elseif (\array_key_exists('examples', $data) && $data['examples'] === null) {
             $object->setExamples(null);
+            unset($data['examples']);
         }
         if (\array_key_exists('format', $data) && $data['format'] !== null) {
             $object->setFormat($data['format']);
+            unset($data['format']);
         }
         elseif (\array_key_exists('format', $data) && $data['format'] === null) {
             $object->setFormat(null);
+            unset($data['format']);
         }
         if (\array_key_exists('contentEncoding', $data) && $data['contentEncoding'] !== null) {
             $object->setContentEncoding($data['contentEncoding']);
+            unset($data['contentEncoding']);
         }
         elseif (\array_key_exists('contentEncoding', $data) && $data['contentEncoding'] === null) {
             $object->setContentEncoding(null);
+            unset($data['contentEncoding']);
         }
         if (\array_key_exists('contentMediaType', $data) && $data['contentMediaType'] !== null) {
             $object->setContentMediaType($data['contentMediaType']);
+            unset($data['contentMediaType']);
         }
         elseif (\array_key_exists('contentMediaType', $data) && $data['contentMediaType'] === null) {
             $object->setContentMediaType(null);
+            unset($data['contentMediaType']);
         }
         if (\array_key_exists('contentSchema', $data) && $data['contentSchema'] !== null) {
             $value_36 = $data['contentSchema'];
@@ -597,9 +709,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $value_36 = $data['contentSchema'];
             }
             $object->setContentSchema($value_36);
+            unset($data['contentSchema']);
         }
         elseif (\array_key_exists('contentSchema', $data) && $data['contentSchema'] === null) {
             $object->setContentSchema(null);
+            unset($data['contentSchema']);
         }
         if (\array_key_exists('definitions', $data) && $data['definitions'] !== null) {
             $values_16 = new \Jane\Component\JsonSchema\JsonSchema\Runtime\JsonObject();
@@ -613,9 +727,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $values_16[$key_6] = $value_38;
             }
             $object->setDefinitions($values_16);
+            unset($data['definitions']);
         }
         elseif (\array_key_exists('definitions', $data) && $data['definitions'] === null) {
             $object->setDefinitions(null);
+            unset($data['definitions']);
         }
         if (\array_key_exists('dependencies', $data) && $data['dependencies'] !== null) {
             $values_17 = new \Jane\Component\JsonSchema\JsonSchema\Runtime\JsonObject();
@@ -635,21 +751,27 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $values_17[$key_7] = $value_40;
             }
             $object->setDependencies($values_17);
+            unset($data['dependencies']);
         }
         elseif (\array_key_exists('dependencies', $data) && $data['dependencies'] === null) {
             $object->setDependencies(null);
+            unset($data['dependencies']);
         }
         if (\array_key_exists('$recursiveAnchor', $data) && $data['$recursiveAnchor'] !== null) {
             $object->setDollarRecursiveAnchor($data['$recursiveAnchor']);
+            unset($data['$recursiveAnchor']);
         }
         elseif (\array_key_exists('$recursiveAnchor', $data) && $data['$recursiveAnchor'] === null) {
             $object->setDollarRecursiveAnchor(null);
+            unset($data['$recursiveAnchor']);
         }
         if (\array_key_exists('$recursiveRef', $data) && $data['$recursiveRef'] !== null) {
             $object->setDollarRecursiveRef($data['$recursiveRef']);
+            unset($data['$recursiveRef']);
         }
         elseif (\array_key_exists('$recursiveRef', $data) && $data['$recursiveRef'] === null) {
             $object->setDollarRecursiveRef(null);
+            unset($data['$recursiveRef']);
         }
         if (\array_key_exists('additionalItems', $data) && $data['additionalItems'] !== null) {
             $value_42 = $data['additionalItems'];
@@ -659,9 +781,16 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
                 $value_42 = $data['additionalItems'];
             }
             $object->setAdditionalItems($value_42);
+            unset($data['additionalItems']);
         }
         elseif (\array_key_exists('additionalItems', $data) && $data['additionalItems'] === null) {
             $object->setAdditionalItems(null);
+            unset($data['additionalItems']);
+        }
+        foreach ($data as $key_8 => $value_43) {
+            if (preg_match('/^x-/', (string) $key_8)) {
+                $object[$key_8] = $value_43;
+            }
         }
         return $object;
     }
@@ -1263,6 +1392,11 @@ class JsonSchemaNormalizer implements DenormalizerInterface, NormalizerInterface
         }
         else {
             $dataArray['additionalItems'] = null;
+        }
+        foreach ($data as $key_8 => $value_43) {
+            if (preg_match('/^x-/', (string) $key_8)) {
+                $dataArray[$key_8] = $value_43;
+            }
         }
         return $dataArray;
     }