+++ src/Component/OpenApi2/Tests/fixtures/x-namespace/expected/Endpoint/Ops/Monitoring/GetTaggedItems.php
@@ -1,0 +1,38 @@
+<?php
+
+namespace Jane\Component\OpenApi2\Tests\Expected\Endpoint\Ops\Monitoring;
+
+class GetTaggedItems extends \Jane\Component\OpenApi2\Tests\Expected\Runtime\Client\BaseEndpoint implements \Jane\Component\OpenApi2\Tests\Expected\Runtime\Client\Endpoint
+{
+    use \Jane\Component\OpenApi2\Tests\Expected\Runtime\Client\EndpointTrait;
+    public function getMethod(): string
+    {
+        return 'GET';
+    }
+    public function getUri(): string
+    {
+        return '/tagged-items';
+    }
+    public function getBody(\Symfony\Component\Serializer\SerializerInterface $serializer, $streamFactory = null): array
+    {
+        return [[], null];
+    }
+    /**
+     * {@inheritdoc}
+     *
+     *
+     * @return null|\Jane\Component\OpenApi2\Tests\Expected\Model\Catalog\TaggedItem
+     */
+    protected function transformResponseBody(\Psr\Http\Message\ResponseInterface $response, \Symfony\Component\Serializer\SerializerInterface $serializer, ?string $contentType = null)
+    {
+        $status = $response->getStatusCode();
+        $body = (string) $response->getBody();
+        if (200 === $status) {
+            return $serializer->deserialize($body, 'Jane\Component\OpenApi2\Tests\Expected\Model\Catalog\TaggedItem', 'json');
+        }
+    }
+    public function getAuthenticationScopes(): array
+    {
+        return [];
+    }
+}